requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
python-dotenv==0.20.0